refactor(macos): extract genre actions & radio from AppModel into AppModel+Genre.swift
Continues the AppModel god-object split. Relocates genre browsing/actions and
the Decade/Mood radio stations — resolvedGenreId, browseGenre/startGenreRadio/
shuffleGenre, probeAvailableMoods/startDecadeRadio/startMoodRadio, loadAlbums/
loadTracks(forGenreId:), pinGenreToHome, and the Mood struct — into a new
`extension AppModel` file.

Pure relocation: the 277 moved lines are byte-identical to what was removed.
AppModel.swift drops 6,147 -> 5,870 lines. The genre/radio stored state
(_genreIdsByName, _genresLoaded, availableMoods) stays on the main class under
a new "Genre / radio state" MARK; the transport controls (pause/resume/stop/
skip/setVolume) physically embedded in the radio MARK range are NOT genre and
stay put.

Four access promotions, all required by the moved methods:
- buildItemsQuery, parseTracksFromItems (private -> internal): shared /Items
  query infra the moodHasTracks / fetchRadioTracks probes route through.
- _genreIdsByName, _genresLoaded (private -> internal): the genre-UUID cache
  the moved resolvedGenreId populates/reads.
The self-contained privates (resolvedGenreId, moodHasTracks, fetchRadioTracks)
move with the domain and stay private.

Verified: clean rebuild (rm -rf macos/.build && swift build) + swift test
(889 tests) green.
